Successfully handling a clone request
That is all it can do but it is pretty cool to see it actuall work.
Requires manually deleting the repository afterwards unfortunately!

We use the ramda library for some functional approaches to handling the
code and then break down and do a bunch of stuff in imperative style. I
still need to work on my functional chops it seems.